core/lock: Add lock timeout warnings
There are currently no timeout warnings for locks in skiboot. We assume
that the lock will eventually become free, which may not always be the
case.

This patch adds timeout warnings for locks. Any lock which spins for more
than 5 seconds will throw a warning and stacktrace for that thread. This is
useful for debugging siturations where a lock which hang, waiting for the
lock to be freed.

Signed-off-by: Matt Brown <matthew.brown.dev@gmail.com>
Reviewed-by: Cyril Bur <cyril.bur@au1.ibm.com>
[stewart: make code match comment, have 5s timeout]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>