pci: Print resetting PHB notice at higher log level
Currently during boot there a long delay while we wait for the PHBs to
be reset and train. During this time, there is no output from skiboot
and the last message doesn't give an indication of what's happening.

This boosts the PHB reset message from info to notice so users can see
what's happening during this long period of waiting.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>