phb4: Workaround phy lockup by doing full PHB reset on retry
For PHB4 it's possible that the phy may end up in a bad state where it
can no longer recieve data. This can manifest as the link not
retraining. A simple PERST will not clear this. The PHB must be
completely reset.

This changes the retry state to CRESET to do this.

This issue may also manifest itself as the link training in a degraded
state (lower speed or narrower width). This patch doesn't attempt to
fix that (will come later).

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>