phb4: Do more retries on link training failures
Currently we only retry once when we have a link training failure.

This changes this to be 3 retries as 1 retry is not giving us enough
reliablity.

This will increase the boot time, especially on systems where we
incorrectly detect a link presence when there really is nothing
present. I'll post a followup patch to optimise our timings to help
mitigate this later.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>