build: CFLAGS compile with -fpie and -fno-pic
When gcc is configured with --enabled-default-pie, sets -fpic which
does not need to be enabled, so remove it. Pass -fpie explicitly as
well just to be careful. This was observed with the powerpc64le cross
compiler for Debian.

This allows the powerpc64le toolchain to compile an almost identical
binary as the powerpc64 toolchain.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>