phb4: Remove retry on electrical link timeout
Currently we retry if we don't detect an electrical link. This is
pointless as all devices should respond in the given time.

This patches removes this retry and just returns OPAL_HARDWARE if we
don't detect an electrical link.

This has the additional benefit of improving boot times on machines
that have badly wired presence detect (ie. says a device is present
when there isn't).

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>