phb4: Only clear some PHB config space registers on errors
Currently on error we clear the entire PHB config space.  This is a
problem as the PCIe Maximum Payload Size (MPS) negotiation may have
already occurred. Clearing MPS in the PHB back to a default of 128
bytes will result an error for a device which already has a larger MPS
configured.

This will manifest itself as error due to a malformed TLP packet. ie.
  phbPblErrorStatus bit 41  = "Malformed TLP error"
This has been seen after kexec on with some adapters.

This fixes the problem by only clearing a subset of registers on a phb
error.

Reported-by: Rob Lippert <rlippert@google.com>
Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>