Revert "phb4: Read PERST signal rather than assuming it's asserted"
This reverts commit b42ff2b904165addf32e77679cebb94a08086966

The original patch assumes that PERST has been asserted well before (>
250ms) we hit here (ie. during hostboot).

In a subesquent patch this will no longer be the case as we need to
assert PERST during PHB reset, which may only be a few milliseconds
before we hit this code.

Hence revert this patch. Go back to the software mechanism using
skip_perst to determine if PERST should be asserted or not. This
allows us to keep the speed optimisation on boot.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>