mambo: Add mambo socket program
This adds a program that can be run inside a mambo simulator in linux
userspace which enables TCP sockets to be proxied in and out of the
simulator to the host.

Unlike mambo bogusnet, it's requires no linux or skiboot specific
drivers/infrastructure to run.

eg.
 Run inside the simulator:
   - to forward host ssh connections to sim ssh server
     ./mambo-socket-proxy -h 10022 -s 22
        Then connect to port 10022 on your host
          ssh -p 10022 localhost
   - to allow http proxy access from inside the sim to local http proxy
     ./mambo-socket-proxy -b proxy.mynetwork -h 3128 -s 3128

Multiple connections are supported.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>