phb4: Use link if degraded
In the recent change:
 3f936bae97 phb4: Retrain link if degraded
We retrain if the link is degraded.  We do 3 retries to get an optimal
link.

Unfortunately if the last retry fails, we mark the PHB as bad and
don't use it. Hence that PHB is lost even though it actually trained
(just degraded).

This fixes the problem by printing an error message (as below) but
still marking the PHB as good.

  [    7.179320404,3] PHB#0005[0:5]: LINK: Link degraded
  [    8.387346665,3] PHB#0005[0:5]: LINK: Link degraded
  [   10.078409137,3] PHB#0005[0:5]: LINK: Link degraded
  [   11.281477269,3] PHB#0005[0:5]: LINK: Link degraded
  [   11.283123885,3] PHB#0005[0:5]: LINK: Degraded but no more retries

Signed-off-by: Michael Neuling <mikey@neuling.org>
Acked-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>