hw/slw.c: Offline code still uses p8 bits
I'm seeing an infinite loop while hot unplugging
a CPU. This is a workaround till we do the right
things for p9. May be a candidate for backporting

The messages I see in an infinite loop are:

[  740.250192896,3] LIBPORE: Core ID = 20 is not within valid range of [0;15]
[  740.250230176,3] SLW: Failed to set spr for CPU 51

When trying to hotunplug core id 20. For now the
patch just skips calling p8_pore* on p9 machines.

Signed-off-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>