pci-iov: free memory across fast-reboot
pci_set_cap needs a callback to free data and we need to call
that when we're doing __pci_reset()

We also need to free pcrf entries.

In the future, __pci_reset() and pci_remove_bus() need to come
together to be one canonical place on how to free a PCI device
rather than the two we have now. This patch *purely* focuses
on the problem of not leaking memory across fast-reboot.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>